+++ lib/kh_calendar.dart
@@ -1,3 +1,4 @@
-library khmer_calendar;
+//library kh_calendar;
 
 export 'src/kh_calendar_widget.dart';
+export 'src/kh_calendar_dialog.dart';